commit final changes
This cruise controller has a timer that counts to 0.0 - 10.0 and displays on LEDs. It also turns a motor to certain speeds, adjusts the speed when weight is placed on the motor slowing it down, and adjusts the speed when weight is taken off to match expected speed values. It has 10 distinct speed values that can be selected from a keypad.